# ui/valset-standardizer-frontend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Java 外部估值表标准化工程，基于 Spring Boot 和 Quartz",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:30066"
    }
  ],
...
  "paths": {
...
    "/api/transfer-sources": {
      "post": {
        "summary": "创建文件来源。",
        "deprecated": false,
        "tags": [
          "TransferSourceController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferSourceUpsertCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultTransferSourceMutationResponse"
                }
              }
            }
          }
        },
        "operationId": "createSource",
        "parameters": []
      },
      "get": {
        "summary": "查询文件来源列表。",
        "deprecated": false,
        "tags": [
          "TransferSourceController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MultiResultTransferSourceViewDTO"
                }
              }
            }
          }
        },
        "operationId": "listSources",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "来源类型",
            "required": false,
            "name": "sourceType"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "来源编码",
            "required": false,
            "name": "sourceCode"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "来源名称",
            "required": false,
            "name": "sourceName"
          },
          {
            "schema": {
              "format": "boolean",
              "type": "boolean"
            },
            "in": "query",
            "description": "是否启用",
            "required": false,
            "example": true,
            "name": "enabled"
          },
          {
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "in": "query",
            "description": "查询上限",
            "required": false,
            "example": 0,
            "name": "limit"
          }
        ]
      }
    },
    "/api/transfer-sources/{sourceId}": {
      "get": {
        "summary": "查询文件来源详情。",
        "deprecated": false,
        "tags": [
          "TransferSourceController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultTransferSourceViewDTO"
                }
              }
            }
          }
        },
        "operationId": "getSource",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "来源主键",
            "required": true,
            "name": "sourceId"
          }
        ]
      },
      "delete": {
        "summary": "删除文件来源。",
        "deprecated": false,
        "tags": [
          "TransferSourceController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultTransferSourceMutationResponse"
                }
              }
            }
          }
        },
        "operationId": "deleteSource",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "来源主键",
            "required": true,
            "name": "sourceId"
          }
        ]
      },
      "put": {
        "summary": "更新文件来源。",
        "deprecated": false,
        "tags": [
          "TransferSourceController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferSourceUpsertCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultTransferSourceMutationResponse"
                }
              }
            }
          }
        },
        "operationId": "updateSource",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "来源主键",
            "required": true,
            "name": "sourceId"
          }
        ]
      }
    },
    "/api/transfer-sources/template-name": {
      "get": {
        "summary": "查询来源对应表单模板名。",
        "deprecated": false,
        "tags": [
          "TransferSourceController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultString"
                }
              }
            }
          }
        },
        "operationId": "getTemplateName_1",
        "parameters": [
          {
            "schema": {
              "format": "enum",
              "type": "string",
              "enum": [
                "LOCAL_DIR",
                "EMAIL",
                "S3",
                "SFTP"
              ]
            },
            "in": "query",
            "description": "来源类型<br/>(See: 文件来源类型。)",
            "required": true,
            "example": "LOCAL_DIR",
            "name": "sourceType"
          }
        ]
      }
    },
    "/api/transfer-sources/{sourceId}/trigger": {
      "post": {
        "summary": "立即触发一次文件来源收取。",
        "deprecated": false,
        "tags": [
          "TransferSourceController"
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultTransferSourceMutationResponse"
                }
              }
            }
          }
        },
        "operationId": "triggerSource",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "来源主键",
            "required": true,
            "name": "sourceId"
          }
        ]
      }
    },
    "/api/transfer-sources/{sourceId}/stop": {
      "post": {
        "summary": "停止文件来源收取。",
        "deprecated": false,
        "tags": [
          "TransferSourceController"
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultTransferSourceMutationResponse"
                }
              }
            }
          }
        },
        "operationId": "stopSource",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "来源主键",
            "required": true,
            "name": "sourceId"
          }
        ]
      }
    },
    "/api/transfer-sources/{sourceId}/checkpoint/processed-mail-ids/clear": {
      "post": {
        "summary": "清空来源已处理邮件标识。",
        "deprecated": false,
        "tags": [
          "TransferSourceController"
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultTransferSourceMutationResponse"
                }
              }
            }
          }
        },
        "operationId": "clearProcessedMailIds",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "来源主键",
            "required": true,
            "name": "sourceId"
          }
        ]
      }
    },
    "/api/transfer-sources/{sourceId}/checkpoints": {
      "get": {
        "summary": "查询来源检查点列表。",
        "deprecated": false,
        "tags": [
          "TransferSourceController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MultiResultTransferSourceCheckpointViewDTO"
                }
              }
            }
          }
        },
        "operationId": "listCheckpoints",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "来源主键",
            "required": true,
            "name": "sourceId"
          },
          {
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "in": "query",
            "description": "查询上限",
            "required": false,
            "example": 0,
            "name": "limit"
          }
        ]
      }
    },
    "/api/transfer-sources/{sourceId}/checkpoint-items": {
      "get": {
        "summary": "查询来源检查点去重记录。",
        "deprecated": false,
        "tags": [
          "TransferSourceController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MultiResultTransferSourceCheckpointItemViewDTO"
                }
              }
            }
          }
        },
        "operationId": "listCheckpointItems",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "来源主键",
            "required": true,
            "name": "sourceId"
          },
          {
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "in": "query",
            "description": "查询上限",
            "required": false,
            "example": 0,
            "name": "limit"
          }
        ]
      }
    },
...
    "/api/transfer-targets/template-name": {
      "get": {
        "summary": "查询投递目标对应表单模板名。",
        "deprecated": false,
        "tags": [
          "TransferTargetController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultString"
                }
              }
            }
          }
        },
        "operationId": "getTemplateName_2",
        "parameters": [
          {
            "schema": {
              "format": "enum",
              "type": "string",
              "enum": [
                "EMAIL",
                "S3",
                "SFTP",
                "FILESYS"
              ]
            },
            "in": "query",
            "description": "目标类型<br/>(See: 文件投递目标类型。)",
            "required": true,
            "example": "EMAIL",
            "name": "targetType"
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TransferSourceMutationResponse": {
        "type": "object",
        "properties": {
          "operation": {
            "format": "string",
            "description": "操作类型。",
            "type": "string"
          },
          "message": {
            "format": "string",
            "description": "提示消息。",
            "type": "string"
          },
          "formTemplateName": {
            "format": "string",
            "description": "关联的表单模板名。",
            "type": "string"
          },
          "source": {
            "description": "文件来源详情。(object)",
            "$ref": "#/components/schemas/TransferSourceViewDTO"
          }
        }
      },
...
      "SingleResultTransferSourceViewDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回的单条数据(object)",
            "$ref": "#/components/schemas/TransferSourceViewDTO"
          }
        }
      },
...
      "SingleResultString": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "format": "string",
            "description": "返回的单条数据",
            "type": "string"
          }
        }
      },
...
      "Object": {
        "type": "object",
        "properties": {
          "mapKey": {
            "description": "A map key.(object)",
            "$ref": "#/components/schemas/Object"
          }
        }
      },
...
      "SingleResultTransferSourceMutationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回的单条数据(object)",
            "$ref": "#/components/schemas/TransferSourceMutationResponse"
          }
        }
      },
...
      "TransferSourceViewDTO": {
        "type": "object",
        "properties": {
          "sourceId": {
            "format": "string",
            "description": "文件来源主键。",
            "type": "string"
          },
          "sourceCode": {
            "format": "string",
            "description": "文件来源编码。",
            "type": "string"
          },
          "sourceName": {
            "format": "string",
            "description": "文件来源名称。",
            "type": "string"
          },
          "sourceType": {
            "format": "string",
            "description": "文件来源类型。",
            "type": "string"
          },
          "formTemplateName": {
            "format": "string",
            "description": "当前使用的来源表单模板名。",
            "type": "string"
          },
          "enabled": {
            "format": "boolean",
            "description": "是否启用。",
            "type": "boolean",
            "example": true
          },
          "pollCron": {
            "format": "string",
            "description": "轮询表达式。",
            "type": "string"
          },
          "ingestStatus": {
            "format": "string",
            "description": "收取状态。",
            "type": "string"
          },
          "ingestBusy": {
            "format": "boolean",
            "description": "是否仍在收取锁定中。",
            "type": "boolean",
            "example": true
          },
          "ingestStartedAt": {
            "format": "string",
            "description": "收取开始时间。",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          },
          "ingestFinishedAt": {
            "format": "string",
            "description": "收取结束时间。",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          },
          "connectionConfig": {
            "description": "连接配置。(object)",
            "$ref": "#/components/schemas/Object"
          },
          "sourceMeta": {
            "description": "来源扩展信息。(object)",
            "$ref": "#/components/schemas/Object"
          },
          "createdAt": {
            "format": "string",
            "description": "创建时间。",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          },
          "updatedAt": {
            "format": "string",
            "description": "修改时间。",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          }
        }
      },
...
      "MultiResultTransferSourceCheckpointViewDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回数据",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransferSourceCheckpointViewDTO"
            }
          }
        }
      },
...
      "TransferSourceUpsertCommand": {
        "type": "object",
        "properties": {
          "sourceId": {
            "format": "string",
            "description": "文件来源主键。",
            "type": "string"
          },
          "sourceCode": {
            "format": "string",
            "description": "文件来源编码。",
            "type": "string"
          },
          "sourceName": {
            "format": "string",
            "description": "文件来源名称。",
            "type": "string"
          },
          "sourceType": {
            "format": "string",
            "description": "文件来源类型。",
            "type": "string"
          },
          "enabled": {
            "format": "boolean",
            "description": "是否启用。",
            "type": "boolean",
            "example": true
          },
          "pollCron": {
            "format": "string",
            "description": "轮询表达式。",
            "type": "string"
          },
          "connectionConfig": {
            "description": "连接配置。(object)",
            "$ref": "#/components/schemas/Object"
          },
          "sourceMeta": {
            "description": "来源扩展信息。(object)",
            "$ref": "#/components/schemas/Object"
          }
        },
        "required": [
          "sourceCode",
          "sourceName",
          "sourceType"
        ]
      },
...
      "TransferSourceCheckpointViewDTO": {
        "type": "object",
        "properties": {
          "checkpointId": {
            "format": "string",
            "description": "检查点主键。",
            "type": "string"
          },
          "sourceId": {
            "format": "string",
            "description": "来源主键。",
            "type": "string"
          },
          "sourceType": {
            "format": "string",
            "description": "来源类型。",
            "type": "string"
          },
          "checkpointKey": {
            "format": "string",
            "description": "检查点键。",
            "type": "string"
          },
          "checkpointValue": {
            "format": "string",
            "description": "检查点值。",
            "type": "string"
          },
          "checkpointMeta": {
            "description": "检查点元数据。(object)",
            "$ref": "#/components/schemas/Object"
          },
          "createdAt": {
            "format": "string",
            "description": "创建时间。",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          },
          "updatedAt": {
            "format": "string",
            "description": "修改时间。",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          }
        }
      },
...
      "MultiResultTransferSourceCheckpointItemViewDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回数据",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransferSourceCheckpointItemViewDTO"
            }
          }
        }
      },
...
      "MultiResultTransferSourceViewDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回数据",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransferSourceViewDTO"
            }
          }
        }
      },
...
      "TransferSourceCheckpointItemViewDTO": {
        "type": "object",
        "properties": {
          "checkpointItemId": {
            "format": "string",
            "description": "去重记录主键。",
            "type": "string"
          },
          "sourceId": {
            "format": "string",
            "description": "来源主键。",
            "type": "string"
          },
          "sourceType": {
            "format": "string",
            "description": "来源类型。",
            "type": "string"
          },
          "itemKey": {
            "format": "string",
            "description": "去重键。",
            "type": "string"
          },
          "itemRef": {
            "format": "string",
            "description": "来源引用标识。",
            "type": "string"
          },
          "itemName": {
            "format": "string",
            "description": "条目名称。",
            "type": "string"
          },
          "itemSize": {
            "format": "int64",
            "description": "条目大小。",
            "type": "number",
            "example": 0
          },
          "itemMimeType": {
            "format": "string",
            "description": "条目 MIME 类型。",
            "type": "string"
          },
          "itemFingerprint": {
            "format": "string",
            "description": "条目指纹。",
            "type": "string"
          },
          "itemMeta": {
            "description": "条目元数据。(object)",
            "$ref": "#/components/schemas/Object"
          },
          "triggerType": {
            "format": "string",
            "description": "触发类型。",
            "type": "string"
          },
          "processedAt": {
            "format": "string",
            "description": "处理时间。",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          },
          "createdAt": {
            "format": "string",
            "description": "创建时间。",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          },
          "updatedAt": {
            "format": "string",
            "description": "修改时间。",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          }
        }
      },
...
    }
  }
}